Simplify model, add docs
Drop the attributes on messages and just keep the marker interfaces. We weren't gaining much, and there is already an attribute for generating the serializer. For value-returning messages, we had to use a generic attribute anyway, so you're not saving much over a generic interface (literally just saving the letter `I` but wasting two extra keystrokes from `[` and `]` (one if you have to add the `:` after the class, though ;)

This makes the `IActorBus` less "magic" and self-evident: it has methods receiving messages that must implement the marker interface you already implement in the messages, so it's less surprising.

Somewhat more controversial and on the "magic" side, is switching to surfacing the `IEventSourced` as a sort of marker interface, in the sense that you don't have to implement it, but rather its implementation is provided automatically for you by a generator. This seems like a more novel approach, which allows you to break free of the inheritance restriction for your actors, without losing a default implementation that is otherwise quite simple in general.

Also added concurrency checks/support in the storage, both for event sourced and plain table storage.