Update sample Streamstone storage with auto-snapshotting
Added automatic snapshotting and quick reloading via snapshots, in a version tolerant way.

The way the sample provider works is:
1. If AutoSnapshot is true, on write, the state will be persisted atomically with Streamstone as a header
2. On Read, in that case, we'll directy attempt to deserialize onto the existing state, the persisted snapshot. If not successfull for any reason (or major.minor assembly version of the type not matching), we'll just fallback to the more expensive read of all events and applying them.

Note that on the subsequent write, the snapshot (if it was version-rejected, say) would be re-created and from that point on, the quick reading would be restored.

This means that to recreate the snapshot, all that's needed is to delete the snapshot.

We use Newtonsoft.Json now since it can deserialize onto an existing instance, which is what we want since the grain can continue to execute other logic on the same grain instance and its associated POCO actor.