Simplify assumption on read of state
In order to account for failures to write to storage, we need to restore the state to its existing state, meaning we'll need to recreate the state instance.

Therefore, our previous snapshot applying to the existing instance is no longer necessary, so we can go back to System.Text.Json.