Leverage Orleans codegen for increased compatiblity
By tweaking the code generator that is public in Orleans, we can emit the necessary attributes and metadata that is regularly expected by Orleans hosting, without needing to extend further via Silo APIs. This also means that we're more compatible with the generated serializers and OOB persistence providers.

The generated grain state is now a plain DTO of the public+private state of the actor class, which is transparently retrieved and set using a memento pattern. For event sourced actors, the state class becomes an event source adapter that invokes the actor automatically. This keeps the streamstone storage sample agnostic to this change.

Since we now generate even the [GenerateSerializer] attributes for messages, we become even more decoupled at the business logic level.