Improve support for validating indirectly used types
Detect lacking partial type in indirect types (i.e. parameters to actor messages). We cannot provide code fixers for these because they have to be reported for an entire compilation (since we need to scan types and usages), which makes the experience suboptimal. But at least we don't fail at run-time.