Split server-only (grains) APIs to our Server package
Related codegen is also moved there, with a detection for the project having the server package installed.

Due to an issue with Orleans codec generation (see https://github.com/dotnet/orleans/issues/9092), end to end tests are failing since state is not being properly serialized for actor messages.

We can only ship this rework once that's fixed. Otherwise, we'll need to move serialization codegen back to the actor/domain project (which unfortunately brings a larger dependency on Orleans, which is undesirable).