Allow reading actor state from storage snapshot
This uncovered a bug in a previous commit which caused snapshot to remain outdated after first save.

The new extension method makes it easy to read the last persisted state of an actor without resorting to Orleans instantiation. This could be quite useful for loading a read-only version of the actor for display, without necessarily incurring RPC and the Orleans overhead unnecessarily.