Rename interfaces project to Abstractions, Server to CloudActors
This aligns better with .NET expectations where the .Abstractions packages are expected to be the one that includes interfaces/contracts only, and implementations are more specific.

Since we have only one implementation possible, we don't add the .Server either to the main CloudActors package which is intended for use in the hosting project.