Proper diagnostics when ProduceReferenceAssembly=true, unsupported
Orleans has a built-in diagnostic to detect this situation (see https://github.com/dotnet/orleans/blob/855a18b6662d04eb64911524561ebeb37240e8a2/src/Orleans.CodeGenerator/Diagnostics/ReferenceAssemblyWithGenerateSerializerDiagnostic.cs), so we detect this property and emit a friendly message.

We also introduce targets to surface that property to the generator as well as defaulting it to `false` (but not preventing users from facing it if they do try to set it to true). This explicitly surfaces this limitation.