Make sure we don't duplicate Orleans-generated code
Since we include the Orleans.CodeGenerator.dll as part of our analyzers, the original source generator there was also being exposed through our assembly, causing Roslyn to invoke essentially the same generator twice. This resulted in duplicate code that made our project unusable.

To avoid this duplication, we ILMerge the Orleans assembly but at the same time remove the generator from it (using Cecil and a simple dotnet file app).