Allow relaxing snapshot version compatibility
Rather than hardcoding a major+minor compatiiblity check between persisted state and current state assembly version, allow relaxing to major version only too.

It's typical in SemVer to use strictly major-only bumps when introducing significant state/API breakage, so if the user uses that, we can relax the requirement on the snapshot reading. This can significantly improve performance when event steam is long and the app is frequently updated but with minor version bumps and no state breaking changes.